- Updated ElectricMeters methods in controller - Added Readings controller with methods - Added Subscribers controller with methods - Added Request/Response models

+++ Backend/ElectricityMeters/ElectricityMeters/Controllers/ElectricMetersController.cs
@@ -31,7 +31,7 @@ namespace ElectricityMeters.Controllers
         {
             if (_dbContext == null || _dbContext.ElectricMeters == null)
             {
-                return Problem("Entity set 'DbContext.ElectricMeters' is null.");
+                return NotFound();
             }
 
             _dbContext.ElectricMeters.Add(electricMeter);
@@ -43,6 +43,11 @@ namespace ElectricityMeters.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> EditElectricMeter(int id, ElectricMeter electricMeter)
         {
+            if (_dbContext == null || _dbContext.ElectricMeters == null)
+            {
+                return NotFound();
+            }
+
             if (id != electricMeter.Id)
             {
                 return BadRequest();
@@ -81,6 +86,13 @@ namespace ElectricityMeters.Controllers
             if (electricMeter == null)
             {
                 return NotFound();
+            }
+
+            // Check if any subscriber is connected to the electric meter
+            var hasConnectedSubscribers = await _dbContext.Subscribers.AnyAsync(s => s.ElectricMeter.Id == id);
+            if (hasConnectedSubscribers)
+            {
+                return BadRequest("Cannot delete the electric meter as there are subscribers connected to it.");
             }
 
             _dbContext.ElectricMeters.Remove(electricMeter);
